backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
PyJWT==2.1.0
python-dotenv==0.19.0
firebase-admin==5.0.0
psycopg2-binary==2.9.1
boto3==1.18.0
